requirements.txt
flask
pytesseract
pillow
opencv-python
numpy
google.generativeai